+++ btb_manager_telegram/settings.py
@@ -1,3 +1,6 @@
 ROOT_PATH = None
 TOKEN = None
 USER_ID = None
+
+TG_UPDATE_BROADCASTED_BEFORE = None
+BTB_UPDATE_BROADCASTED_BEFORE = None
